refactor: move progress indicator widget
