+++ lib/home.dart
@@ -6,11 +6,11 @@ import 'package:lux/const/const.dart';
 import 'package:lux/core_manager.dart' hide ProxyMode;
 import 'package:lux/dashboard.dart';
 import 'package:lux/process_manager.dart';
-import 'package:lux/progress_indicator.dart';
 import 'package:lux/tr.dart';
 import 'package:lux/tray.dart';
 import 'package:lux/utils.dart';
 import 'package:lux/webview_dashboard.dart';
+import 'package:lux/widget/progress_indicator.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:path/path.dart' as path;
 import 'package:tray_manager/tray_manager.dart';
